Fixed the issue with albums on imgur.com
The original filter, 
imgur.com##div#under-image

was too aggressive. It got rid of the comments, as you would hope, but also got rid of the link at the bottom of longer albums that is needed to load the rest of the album. I have replaced it with three filters which have the desired effect of removing comments without removing this necessary feature:

imgur.com###caption_textarea
imgur.com##.relative.action-bar
imgur.com##.small.textbox.comments-info